Make multiple origin processors incremental
Multiple origin processors can aggregate multiple source
elements into one output element or create warnings/errors
for such combinations of elements. They need to be provided
with all their elements of interest in the next compilation
when any of them has changed.

We handle this with a very simple approach: If a processor
registers itself for "*" (a.k.a all root elements), then we
always do a full recompile. If a processor only registers itself
for specific annotations, then we treat all elements with those
annotations as dirty on every compilation. This means that multiple
origin processors create small non-incremental clusters of code.